QueryOptimizationModule created

+++ profiler/views.py
@@ -1,5 +1,4 @@
 import asyncio
-import json
 from django.http import JsonResponse, HttpResponse
 from profiler.models import DatabaseProfile
 from profiler.serializers import DatabaseProfileSerializer
@@ -11,15 +10,14 @@ from common.logging import logerror, Timing
 class IndexView(HybridJsonView):
     def get_content(self, request):
         try:
-            db_profiles = DatabaseProfileSerializer(
-                DatabaseProfile.objects.filter(completion_progress=100),
-                many=True,
-                children=['teachers', 'classes', 'students']
-            ).data
+            db_profiles = DatabaseProfile.objects.filter(completion_progress=100)
+
         except DatabaseProfile.DoesNotExist:
             db_profiles = []
 
-        return {'db_profiles': db_profiles, 'title': 'Profiler'}
+        serializer = DatabaseProfileSerializer(db_profiles, many=True)
+
+        return {'db_profiles': serializer.data, 'title': 'Profiler'}
 
 
 @request_validation(['POST'], expected_body=DatabaseProfileSerializer)
@@ -30,10 +28,10 @@ def create_database_profile(request, body):
     except Exception as e:
         return handle_database_exceptions(e)
 
-    # Running save_sets inside of a asyncio coroutine makes the expensive save_sets() run much faster (because
-    # asyncio coroutines are lightweight and inherently much faster) and also runs it asynchronously in the background
-    # so the simple new_profile can be returned with its db_profile_id and the front end can make subsequent calls to
-    # check on the completion_progress of new_profile while save_sets() is running.
+    # Running save_sets inside of a asyncio coroutine makes the expensive save_sets() run much faster (not sure why)
+    # and also runs it asynchronously in the background so the simple new_profile can be returned with its
+    # db_profile_id and the front end can make subsequent calls to check on the completion_progress of new_profile
+    # while save_sets() is running.
     loop = asyncio.new_event_loop()
     async def save_sets():
         def try_save_sets():
@@ -49,11 +47,7 @@ def create_database_profile(request, body):
     # TODO: Figure out how to avoid this
     loop.close()
 
-    response = JsonResponse({
-        'db_profile': DatabaseProfileSerializer(new_profile).data
-    })
-
-    return response
+    return JsonResponse({'db_profile': DatabaseProfileSerializer(new_profile).data})
 
 
 @request_validation(['GET'])
@@ -62,16 +56,62 @@ def check_progress(request, db_profile_id):
         db_profile = DatabaseProfile.objects.get(db_profile_id=db_profile_id)
     except DatabaseProfile.DoesNotExist:
         return HttpResponse(status=404)
-    children = ['teachers', 'classes', 'students'] if db_profile.completion_progress == 100 else None
-    serialized = DatabaseProfileSerializer(db_profile, children=children).data
-    return JsonResponse({'db_profile': serialized})
+    return JsonResponse({'db_profile': DatabaseProfileSerializer(db_profile).data})
 
 
-@request_validation(['DELETE'])
-def delete_database_profile(request, db_profile_id):
-    try:
-        db_profile = DatabaseProfile.objects.get(db_profile_id=db_profile_id)
-    except DatabaseProfile.DoesNotExist:
-        return HttpResponse(status=404)
-    db_profile.delete()
-    return HttpResponse(status=200)
+@request_validation(['GET', 'DELETE'])
+def database_profile(request, db_profile_id):
+    if request.method == 'DELETE':
+        try:
+            db_profile = DatabaseProfile.objects.get(db_profile_id=db_profile_id)
+        except DatabaseProfile.DoesNotExist:
+            return HttpResponse(status=404)
+        db_profile.delete()
+        return HttpResponse(status=200)
+
+    if request.method == 'GET':
+        teacher_levels = request.GET.get('teacher_levels')
+        class_levels = request.GET.get('class_levels')
+        student_levels = request.GET.get('student_levels')
+        prefetch_related = request.GET.get('prefetch_related')
+
+        def generate_prefetch_command(commands):
+            return lambda levels: "__".join(commands[:levels])
+
+        prefetch_commands = {
+            'teacher': generate_prefetch_command(['teacher_set', 'classes', 'student_set']),
+            'class': generate_prefetch_command(['class_set', 'student_set', 'classes']),
+            'student': generate_prefetch_command(['student_set', 'classes', 'student_set'])
+        }
+
+        prefetch_related_args = ()
+        if teacher_levels:
+            teacher_levels = int(teacher_levels)
+            if prefetch_related == "true" and teacher_levels != 0:
+                prefetch_related_args = prefetch_related_args + (prefetch_commands['teacher'](teacher_levels),)
+        if class_levels:
+            class_levels = int(class_levels)
+            if prefetch_related == "true" and class_levels != 0:
+                prefetch_related_args = prefetch_related_args + (prefetch_commands['class'](class_levels),)
+        if student_levels:
+            student_levels = int(student_levels)
+            if prefetch_related == "true" and student_levels != 0:
+                prefetch_related_args = prefetch_related_args + (prefetch_commands['student'](student_levels),)
+
+        try:
+            db_profile = DatabaseProfile.objects\
+                .filter(db_profile_id=db_profile_id)\
+                .prefetch_related(*prefetch_related_args)
+        except DatabaseProfile.DoesNotExist:
+            return HttpResponse(status=404)
+
+        serializer = DatabaseProfileSerializer(
+            db_profile,
+            many=True,
+            student_levels=student_levels,
+            class_levels=class_levels,
+            teacher_levels=teacher_levels
+        )
+
+        return JsonResponse({'db_profile': serializer.data})
+
